:ambulance: fix build #256
